Fix pylint(no-member) when accessing resource.id
Pylint currently reports `E1101: Instance of 'Bucket' has no 'id' member (no-member)` on lines in Pulumi Python programs like:

```python
pulumi.export('bucket_name', bucket.id)
```

Here's a description of this message from http://pylint-messages.wikidot.com/messages:e1101:

> Used when an object (variable, function, …) is accessed for a non-existent member.
>
> False positives: This message may report object members that are created dynamically, but exist at the time they are accessed.

This is a false positive because `id` isn't set in the initializer method, it's set later in `register_resource` and Pylint isn't able to figure this out statically.

This change refactors `register_resource` so that `id` (and `urn`) can be assigned in the resource's initializer, so that Pylint can see it being assigned, and does the same for `read_resource`.