Refactor the backend package.
The primary goal of this refactoring is to simplify the interface
that a Pulumi backend must implement in order to provide better clarity
and separation of concerns. The new interface is centered around the set
of operations necessary for persistence rather than the set of behaviors
required of a stack (e.g. statefile storage vs. running an update or
preview). The former functionality is captured by a new interface,
`backend.Client`, and the latter is captured in the concrete type
`cli.Backend.` Implementers of `backend.Client` should not perform
console I/O--that is the purview of `cli.Backend`.

These changes lay the groundwork for concurrency control, project/org
support, etc. in the filestate backend.