Use Python properties for resource outputs
This enables docstrings to show up when hovering over properties in PyCharm and VS Code (w/ Pylance).

The property values are stored in and retrieved from the object's `__dict__`, which allows existing things like `export("res", res)` to continue to work the same as before, and aligns the input/output types to store values in `__dict__` the same way.