Changes to input/output types
Two main changes:

1. Previously, if an @output_type class didn't have an __init__ method,
one would be added by the decorator that accepted a single dict argument
representing the output properties. This commit makes it so that instead
of accepting a single dict argument, it accepts individual arguments for
each property based on annotations defined on the class. Classes
decorated with @input_type now do the same thing.

2. Values in input/output types are now stored in `self.__dict__` as
snake_case Python names, rather than being stored in `self._values` as
camelCase Pulumi names, to be consistent with how values are stored for
resources. This allows these types to work the same as resources when
passed to `pulumi.export`. There are two exceptions for output types:
(a) if the class is a subclass of dict, the values are stored in the
dict itself (rather than __dict__), and (b) if the output class has a
`_translate_property` method, the names will be the translated names
from calling that method.