Fixed issue #5014 (Missing Pulumi console tags on Windows)
The existing `path.Join(gitRoot, "..")` function appears to have unexpected behavior on Windows. Instead of appending `..` and returning gitRoot's parent directory,
it replaces the whole string with a single dot. The fix changes the code to instead use `filepath.Dir(gitRoot)`, which correctly returns the parent directory.

This seems to be the only instance in which `path.Join()` is used with `..` in the entire repository. There are some usages of a function called `filepath.Join()`
with dots, but they work correctly.