Add support for getResource to the Python SDK.
Just what it says on the tin.

These changes take advantage of the fact that Python invokes are
synchronous to fetch the resource state inside of its constructor. The
SDK code generator will be updated to use the new `urn` parameter inside
of each module's implementation of `ResourceModule.construct`.

Part of #2430.