Proxy missing Python dict entries during preview
This is a hypothetical approach to fixing #4040. I'm sure there
are fatal flaws in this, however, ever since encountering the bug
this is what I assumed we would do.

The idea is to catch missing keys (and properties, due to typed
dicts), and substitute fake outputs that never resolve to known
values during previews. This enusres that keys that are only
populated during an apply show up inside of the dicts as expected.

The two flaws I see immediately are that the RTTI for the returned
object won't actually be a dict, so type testing may not work,
in addition to, I suppose, the false negative for keys that truly
won't ever get populated, and deferring the knowledge of that until
update rather than seeing them during applies.