Avoid accidental mutation of transform arguments
The contract for `transformations` functions is that they should not modify the `props` and `opts` arguments, but should instead return a new `props` and/or `opts`. However, this isn't enforced, and if user code *does* mutate the objects passed in - this can lead to unexpected results.

With this change, we shallow clone the `props` and `opts` passed to user code to avoid these issues.  This is only a shllow clone, as the objects themselves are complex (and can have internal cycles).  This does still mean users could theoretically mutate state in an unsafe way - but the common case is still protected against.

Fixes #5318.