Allow VERSION to be set explicitly
When building for alpine aports we download the pulumi source from a
github archive (e.g.
https://github.com/pulumi/pulumi/archive/v2.16.0.tar.gz) but this
doesn't include the .git information to allow scripts/get-version to
work.

This patch, which is currently applied on top of the base pulumi
code as part of the alpine build (see [1]), simply allows the alpine
build script to explicitly set the version used. In the alpine build
script this always matches the source code archive version to download.

[1] https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/16281