[codegen/pcl] Add a urnName resource option.
This resource option controls the "name" portion of the resource's URN.
It is intended for use by the import code generator, which may need to
generate code for multiple resources that share the "name" portion of
the URN, but must still generate unique names for the PCL resource
declarations.

This is the backend half of the fix for #6032. The frontend half of the
fix will ensure that names for PCL resource declarations are unique.