Fix #5577
This change fixes the problem where a missing asset or archive
will silently fail. This is because the engine returns an RPC error
which is ignored by (some) language SDKs. Instead of relying on the
SDK to catch and deal with this -- which it should, of course -- this
change issues an error straight from the evaluation source.

I've verified locally that this addresses the bug. Instead of a silent
resource going missing, I see an error:

     ├─ aws:lambda:Function    mylambdaDoesNotExist              1 error

    Diagnostics:
      aws:lambda:Function (mylambdaDoesNotExist):
        error: problem with resource properties: failed to compute archive
            hash: couldn't read archive path './DoesNotExist':
                stat ./DoesNotExist: no such file or directory

I'll write some regression tests too.

I left some questions in the issue that still need feedback. Namely,

1) What do we do with all of the other RPC failure modes in here?

2) What do we do with the SDKs themselves, and the fact that they
   are apparently masking errors by ignoring RPC failures?

It's quite possible we don't want to go down the path I have here,
and instead rely on the SDKs, however to the extent that an error is
always a fatal error -- like unmarshalling a RegisterResource call --
it seems wise to catch it at the source and minimize chance of mistakes.