Add a regression test
This tests the missing asset case.

There is still one outstanding issue: not sure why the Pulumi
CLI returns a 0 exit code, given that errors have been issued.
I vaguely recall that we stopped returning non-0 error codes
purely based on the presence of error messages, and in favor of
bailing, at some point. Likely this means the original fix needs
to do "more" to actually trigger an error (or we should be exiting
with non-0 exits when errors are present!)