Update python binary lookup logic to follow symlinks
`python3.exe` is a symlink for `python.exe` on some Python installations for Windows. Running `python3 -m venv venv` fails with `Error: [WinError 2] The system cannot find the file specified`. However `python -m venv venv` succeeds as one would expect. To workaround, if we find `python3` on the PATH, we'll check to see if it's a symlink, and if so, follow the symlink.